add get_size
